backend/requirements.txt
Flask
mysql-connector-python
opencv-contrib-python
numpy
Pillow
